fix: Refactor fixtures
