+++ tests/to_refactor/fixtures/simulation_analysis_fixtures.py
